Move close nav on document click function to separate file
This allows us to simplify header.js and hopefully makes the code easier to maintain
